+++ react/README.md
@@ -1332,7 +1332,7 @@ Different projects choose different testing tradeoffs based on how often compone
 ```js
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from AppRedux;
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
